Extract functions from views to separate files
