+++ concent_api/core/exceptions.py
@@ -1,3 +1,7 @@
 
 class UnexpectedResponse(Exception):
     pass
+
+
+class Http400(Exception):
+    pass
